Avoid adding an attribute.

+++ src/data/array_interface.cu
@@ -5,7 +5,7 @@
 #include "../common/common.h"
 
 namespace xgboost {
-void ArrayInterfaceHandler::SyncCudaStream(ptrdiff_t stream) {
+void ArrayInterfaceHandler::SyncCudaStream(int64_t stream) {
   dh::safe_cuda(cudaStreamSynchronize(reinterpret_cast<cudaStream_t>(stream)));
 }
 }  // namespace xgboost
